Move common docker and ec2 code to common package
