+++ service/common/certauth.go
@@ -1,4 +1,4 @@
-package docker
+package common
 
 import (
 	"crypto/rand"
