+++ service/common/install.go
@@ -1,0 +1,6 @@
+package common
+
+const (
+	ReleaseUrl         = "http://latestbuilds.service.couchbase.com/builds/releases/"
+	BuildUrl           = "http://latestbuilds.service.couchbase.com/builds/latestbuilds/couchbase-server/"
+)
